Refactor the code to make the chess board independent from the screen
Initially, I wanted to take advantage of pygame Sprites to make the
chess board independent from the screen by making it a combination of
Sprites - the board itself and the pieces. The main idea was that, since
I wanted to make the board independent from the screen, I need to make
the board and the pieces render on a different Surface and then blit
it on the screen, but I was not sure how I can combine multiple Sprites
to form one separate object that can be drawn and updated with
interactions.

The main complication that I encountered, but could not find a
good solution to, is that finding where on the board a player has
clicked is not straight-forward at all. I had 2 solutions:

1) I can make every chess board square a Sprite object and then use
rect.collidepoint() to find if a player clicked on a particular square
and, consequently, a piece that might or not be on that square. This
would, ideally, allow us to work with object directly and not check for
any coordinate-related stuff. The the issue with this solution was that
I had to refactor most of the code and make squares somehow accomodate
pieces on them and then everything becomes more complicated since now we
have the board, the squares, and the pieces. Do we render the board
using squares and not a static background picture (so no board Sprite
whatsoever)? Should we store if a square is a "potential move" square
for the active piece inside of that square? This solution makes
everything much more difficult and is completely different from what I
had before, so I decided not to do it this way (at least, for the time
being, since I like the idea and it might prove to be a more elegant
solution).

2) I can keep using the mouse coordinates to find the position on the
board that a player clicked on by simply taking into account the offset
of the board from the screen. In addition to this I can use
rect.collidepoint() to check if a player actually clicked on the board
and not somewhere else on the screen. This solution seemed more
appropriate for such a simple game/application, so I decided to go with
it, at least for now.

Other changes include:

1) I changed how the board background is drawn on the screen, now the
tiles are drawn on a Surface only once in the beginning and then that
Surface is reused when needed.

2) I changed the chess pieces to be subclasses of pygame.sprite.Sprite,
which should, potentially, make inter-object interactions easier to
implement.

3) Following the previous change, I changed how images are loaded for
the pieces, which has resulted in less code needed to do that.