Add visual improvements, make game turn-based, and small changes
I changed the background of the application, added blur to the chess
board to indicate that the game is not in progress (right now you can
start the game by simply clicking on the blurred board).

The game is finally getting playable, I added game turns for sides, so
that only pieces of one colour can be moved during each turn.

Other small changes include adding internal board coordinates for the
pieces that are updates on moves, a slightly different board border, and
minor changes to the structure of the functions.