Fix bugs with finding potential piece positions and add comments
I fixed 2 major bugs with the function related to finding either the
positions that will be under attack on the next enemies turn or
potential moves overal.

Bug 1:
Pawns would not check if there are pieces in front of them, which would
result in a possibility of jumping over pieces on the first move.

Bug 2:
When calculating possible attack positions of the enemy pieces I
initially assumed I can simply find potential move positions for any
piece, which proved wrong, since pawns move and attack differently.