Fix major bug related to encapsulation of a method
I fixed a major bug with a method does_move_expose_king(), which would
crash the application during the update of the checkmate status. The bug
was essentially me forgetting to preserve the values related to the last
turn information inside does_move_expose_king(), which would result in
update_is_king_in_checkmate() returning to different values when trying
to revert the last move in its scope.